+++ CHANGELOG.md
@@ -22,5 +22,9 @@
 - Fixed bug, now both tracking and event data are normalized in direction
 - Fixed unexpected behaviour, all date related objects are now datetime objects
 
+## V0.3.0 (date)
+
+- Added way to save Match objects, and to load saved Match objects
+- Fixed bug in opta event data, own goals are now parsed as seperate event type